api/app/requirements.txt
fastapi == 0.104.0
gunicorn == 21.2.0
uvicorn == 0.23.2
mysql-connector-python == 8.1.0
python-multipart == 0.0.6
pydub == 0.25.1